work on renaming cookies service to be more specific, ready for new journey cookie service
