+++ src/Dfe.EarlyYearsQualification.Web/Views/Shared/Components/CookiesBanner/Default.cshtml
@@ -1,6 +1,6 @@
-@using Dfe.EarlyYearsQualification.Web.Services.CookieService
+@using Dfe.EarlyYearsQualification.Web.Services.CookiesPreferenceService
 @model Dfe.EarlyYearsQualification.Web.Models.Content.CookiesBannerModel;
-@inject ICookieService cookieService
+@inject ICookiesPreferenceService cookieService
 
 @{
     var cookie = cookieService.GetCookie();
